Make grep just basic without using any env-variables.
That is, don't use GREP_COLOR or GREP_OPTIONS, otherwise color-codes will be
printed in the html-file, which don't show nice :)

This might be too much modification... maybe only a few grep-commands would be
enough to disable the use of these environment variables. However, this works :)
and I don't see any regression.